Move observe_children context-manager to TestCase class to make it available for functional and integration tests
