+++ opengever/testing/integration_test_case.py
@@ -325,25 +325,6 @@ class IntegrationTestCase(TestCase):
         else:
             raise ValueError('Unsupport lookup entry type {!r}'.format(type_))
 
-    @contextmanager
-    def observe_children(self, obj, check_security=True):
-        """Observe the children of an object for testing that children were
-        added or removed within the context manager.
-        """
-        if check_security:
-            def allowed(obj):
-                return getSecurityManager().checkPermission(
-                    'Access contents information', obj)
-        else:
-            def allowed(obj):
-                return True
-
-        children = {'before': filter(allowed, obj.objectValues())}
-        yield children
-        children['after'] = filter(allowed, obj.objectValues())
-        children['added'] = set(children['after']) - set(children['before'])
-        children['removed'] = set(children['before']) - set(children['after'])
-
     def get_catalog_indexdata(self, obj):
         """Return the catalog index data for an object as dict.
         """
